Fixes tests that were failing under Node.js 14.x
GitHub has updated the version of Node.js in all hosted virtual environments
from 12.x to 14.x (actions/virtual-environments#1953).

With this change, a few tests that create empty files on disk
stopped working because `writeFileSync` no longer accepts `Object`
as the data argument. The solution was to replace the empty object literal
with an empty string.