Runs the build when pushing to `master` or a PR branch
Previously, pushing to a PR branch would cause the build
to run twice: once in response to the `push` event and once
in response to the `pull_request` event.

This would often cause one of the builds to fail, due to a
race condition when uploading the code coverage report to
Coveralls.

This commit prevents this from happening by running the build
only on pushes to the `master` branch, as well as to any branch
associated to a PR.