Consolidates the parsing of the `script-path` and `csproj-path` input parameters
This commit introduces a new type to represent the input file called `File`:

```typescript
{
  type: string,
  path: string
}
```

`File` is a discriminated union type that can either be a `ScriptFile`
or a `ProjectFile`. The `path` field is set to either the value of the
`script-path` or `csproj-path` input parameter, based on what is provided
by the user.

This new abstraction makes is clear that only one type of file
can be executed at any one time by the Cake Action. Also, it allows us to
more clearly express that the `csproj-path` input parameter takes precedence
over the `script-path` parameter.