Upgrades to TypeScript 5 and ESLint 9
As part of the upgrade to ESLint 9, we migrated the ESLint configuration
file to the new flat config format. With this change, the `__tests__`
directory has been moved to the `src` directory in order to properly
lint the test files.

We also upgraded other related dev dependencies, such as Jest and
Coveralls.

See also:

- https://eslint.org/docs/latest/use/configure/migration-guide