Skips uninstalling the Cake.Tool when the required version is already installed (#62)
When installing the Cake.Tool in the tools directory, the name of the
subdirectory that contains the package metadata is the package's ID
in lower case.

Unaware of that fact, we were checking for the presence of that
subdirectory using the unmodified package ID ("Cake.Tool") when
determining whether the required version of the tool was already
installed.

This caused the tool to always be reinstalled, since we were looking for
the package metadata in the wrong place.

This commit fixes the problem by normalizing the package ID when
constructing the path to the package metadata.