Upgrades to Jest 30
There's currently a [bug][1] in `ts-jest` that triggers a compiler warning about
the `isolatedModules` options having to be set to `true` when using hybrid modules.

Reading up on a [related issue][2], it seems that `ts-jest` had itself an option called
`isolatedModules`, which behaved differently than the one [defined by the TypeScript
compiler][3].

Right now, the only way to silence the warning is to set that option in the TypeScript
compiler, which doesn't affect the runtime behavior of the tests like the one in
`ts-jest` does.

[1]: https://github.com/kulshekhar/ts-jest/issues/5013
[2]: https://github.com/kulshekhar/ts-jest/issues/4859
[3]: https://www.typescriptlang.org/tsconfig/#isolatedModules