Correctly verifies the error handling in the entry-point
Previously, we mocked the `cake.runScript` function when verifying the
error handling in the action's entry-point.

However, as the entry-point evolved during the years, the `cake.runScript`
function no longer runs during those tests, due to the fact that the
`action.getInputs` function returns `undefined`.

This commit addresses the problem by mocking `action.getInputs` instead,
which is in line with other tests for the same module.