Handles the case when an unknown error type is thrown in the entry-point
With the upgrade to Jest 30 in 5c27d53, we enabled the [`isolatedModules`][1]
option in the TypeScript compiler in order to silence a warning.

This change uncovered an untested branch in the entry-point, where error
types other than `Error` or `string` weren't being handled.

For the sake of correctness, this commit handles that case, as well.

[1]: https://www.typescriptlang.org/tsconfig/#isolatedModules